library/axi_hdmi_tx: Use ring counter for FIFO
The original binary counter was replaced with a ring counter. In this
way there is no need for binary-to-gray and gray-to-binary conversion,
because the ring counter is much less prone to metastable states.

The fifo module is changed to fifo_mem1.

This architecture was inspired by: https://ieeexplore.ieee.org/document/4208997

Signed-off-by: Daniel Matyas <matyasdaniel004@gmail.com>