library/axi_hdmi_tx: Add oh2b into HDMI TX
This addition is not correct logically. It was added only to compare the
implemenatation results (power, utilization, timing) of this approach
with other architectures.

TODO: Add a system, possibly making use of programmable delay synchronizers
from the Jex FIFO implementation to make the pipelined version of oh2b
conversion compatible with HDMI TX. It should start the conversion 9
clock cycles before the result is needed. Another solution is to use the
binary results only when a ready signal is asserted.

Signed-off-by: Daniel Matyas <matyasdaniel004@gmail.com>